requirements.txt
pandas
scikit-learn
torch
transformers
nltk
jsonlines
sentence-transformers
python-dotenv
tensorflow
tensorflow-hub
huggingface-hub
openai
accelerate